+++ CHANGELOG.rst
@@ -157,6 +157,8 @@ Added
   `#485 <https://github.com/openego/eGon-data/issues/485>`_
 * Include time-dependent coefficient of performance for heat pumps
   `#532 <https://github.com/openego/eGon-data/issues/532>`_
+* Limit number of parallel processes per task
+  `#265 <https://github.com/openego/eGon-data/issues/265>`_
 
 .. _PR #159: https://github.com/openego/eGon-data/pull/159